feat: sheets view/edit split + 3-dot card menu + sync icon fix
- clicking a config card navigates to a new read-only view page ([configId]/page.tsx)
- view page shows label, metadata row (type/tab/status/last synced), read-only mapping table, and danger zone
- mapping table columns wrap instead of truncate, rows vertically centered
- Edit button (top right of view page) navigates to [configId]/edit/page.tsx
- edit page moved to [configId]/edit/page.tsx, back/cancel return to view page
- danger zone removed from edit page, lives on view page only
- sheets index: removed SplitButton export, replaced with borderless 3-dot menu per card
- 3-dot menu groups: export JSON/CSV | edit | delete (with confirm modal inline)
- sync button no longer renders IconSync when loading (spinner already shown)
- added IconEdit, IconTrash, IconDotsVertical, IconExport to Icons.tsx
