+++ frontend/app/dashboard/[tournamentId]/sheets/[configId]/page.tsx
@@ -1,48 +1,38 @@
 "use client";
 
-import { useEffect, useState, useRef, useCallback } from "react";
+import { useEffect, useState } from "react";
 import { useRouter, useParams } from "next/navigation";
-import { membershipsApi, sheetsApi, ColumnMapping, SheetConfig } from "@/lib/api";
-import {
-  MappingRow,
-  MappingsExport,
-  ImportSummary,
-  parseMappingsJson,
-  parseMappingsCsv,
-  applyImport,
-} from "@/lib/importMappings";
-import { SplitButton } from "@/components/ui/SplitButton";
+import { membershipsApi, sheetsApi, SheetConfig } from "@/lib/api";
 import { Button } from "@/components/ui/Button";
 import { Banner } from "@/components/ui/Banner";
-import { ImportSummaryModal } from "@/components/ui/ImportSummaryModal";
 import { FieldLabel } from "@/components/ui/FieldLabel";
-import { IconArrowLeft, IconWarning } from "@/components/ui/Icons";
+import { IconArrowLeft, IconEdit, IconWarning } from "@/components/ui/Icons";
 
 // ─── Constants ────────────────────────────────────────────────────────────────
 
-const SHEET_TYPES = [
-  { value: "interest",     label: "Interest Form" },
-  { value: "confirmation", label: "Confirmation Form" },
-  { value: "events",       label: "Events" },
-];
+const SHEET_TYPE_LABELS: Record<string, string> = {
+  interest:     "Interest Form",
+  confirmation: "Confirmation Form",
+  events:       "Events",
+};
 
 const KNOWN_FIELDS_LABELS: Record<string, string> = {
   "__ignore__":          "Ignore",
-  "first_name":          "First Name",
-  "last_name":           "Last Name",
-  "email":               "Email",
-  "phone":               "Phone",
-  "shirt_size":          "Shirt Size",
-  "dietary_restriction": "Dietary Restriction",
-  "university":          "University",
-  "major":               "Major",
-  "employer":            "Employer",
-  "role_preference":     "Role Preference",
-  "event_preference":    "Event Preference",
-  "availability":        "Availability",
-  "lunch_order":         "Lunch Order",
-  "notes":               "Notes",
-  "extra_data":          "Extra Data",
+  first_name:            "First Name",
+  last_name:             "Last Name",
+  email:                 "Email",
+  phone:                 "Phone",
+  shirt_size:            "Shirt Size",
+  dietary_restriction:   "Dietary Restriction",
+  university:            "University",
+  major:                 "Major",
+  employer:              "Employer",
+  role_preference:       "Role Preference",
+  event_preference:      "Event Preference",
+  availability:          "Availability",
+  lunch_order:           "Lunch Order",
+  notes:                 "Notes",
+  extra_data:            "Extra Data",
 };
 
 const TYPE_LABELS: Record<string, string> = {
@@ -55,30 +45,13 @@ const TYPE_LABELS: Record<string, string> = {
   category_events: "Category Events",
 };
 
-const selectStyle: React.CSSProperties = {
-  height: "36px", padding: "0 10px",
-  border: "1px solid var(--color-border)", borderRadius: "var(--radius-sm)",
-  fontFamily: "var(--font-sans)", fontSize: "12px",
-  color: "var(--color-text-primary)", background: "var(--color-bg)",
-  outline: "none", cursor: "pointer",
-};
-
-// ─── Row state ────────────────────────────────────────────────────────────────
-
-type RowState = "same" | "new" | "removed" | "changed";
-
-interface RichMappingRow extends MappingRow {
-  state: RowState;
-  /** Original saved values — used to detect if user reverted a change */
-  original: MappingRow;
+function fmtDateTime(iso: string) {
+  const normalized = iso.endsWith("Z") || iso.includes("+") ? iso : iso + "Z";
+  return new Date(normalized).toLocaleString("en-US", {
+    month: "short", day: "numeric", year: "numeric",
+    hour: "numeric", minute: "2-digit", timeZoneName: "short",
+  });
 }
-
-const ROW_COLORS: Record<RowState, { bg: string; border: string } | null> = {
-  same:    null,
-  new:     { bg: "#F0FDF4", border: "#86EFAC" },
-  removed: { bg: "#FFF5F5", border: "#FCA5A5" },
-  changed: { bg: "#FEFCE8", border: "#FDE047" },
-};
 
 // ─── Delete Confirmation Modals ───────────────────────────────────────────────
 
@@ -129,6 +102,15 @@ function DeleteMembershipsModal({
   const [typed, setTyped] = useState("");
   const confirmed = typed === "DELETE";
 
+  const inputStyle: React.CSSProperties = {
+    height: "36px", padding: "0 10px", width: "100%",
+    border: `1px solid ${typed && !confirmed ? "var(--color-danger)" : "var(--color-border)"}`,
+    borderRadius: "var(--radius-sm)",
+    fontFamily: "var(--font-mono)", fontSize: "13px",
+    color: "var(--color-text-primary)", background: "var(--color-bg)",
+    outline: "none",
+  };
+
   return (
     <div
       style={{ position: "fixed", inset: 0, background: "rgba(0,0,0,0.5)", zIndex: 200, display: "flex", alignItems: "center", justifyContent: "center" }}
@@ -167,12 +149,7 @@ function DeleteMembershipsModal({
             onChange={(e) => setTyped(e.target.value)}
             placeholder="DELETE"
             autoFocus
-            style={{
-              ...selectStyle,
-              width: "100%",
-              borderColor: typed && !confirmed ? "var(--color-danger)" : "var(--color-border)",
-              fontFamily: "var(--font-mono)",
-            }}
+            style={inputStyle}
           />
         </div>
 
@@ -194,180 +171,92 @@ function DeleteMembershipsModal({
   );
 }
 
-// ─── Mapping Table Row ────────────────────────────────────────────────────────
+// ─── Read-only Mapping Table Row ──────────────────────────────────────────────
 
-function MappingTableRow({
-  row,
-  knownFields,
-  validTypes,
-  onChange,
+function ReadOnlyMappingRow({
+  header,
+  field,
+  type,
+  row_key,
+  extra_key,
   isLast,
 }: {
-  row: RichMappingRow;
-  knownFields: string[];
-  validTypes: string[];
-  onChange: (patch: Partial<MappingRow>) => void;
+  header: string;
+  field: string;
+  type: string;
+  row_key?: string;
+  extra_key?: string;
   isLast: boolean;
 }) {
-  const isRemoved    = row.state === "removed";
-  const needsRowKey  = row.type === "matrix_row";
-  const needsExtraKey = row.field === "extra_data";
-  const colors       = ROW_COLORS[row.state];
-
-  function handleFieldChange(field: string) {
-    let type = row.type;
-    if (field === "__ignore__")                                              type = "ignore";
-    else if (field === "availability")                                       type = "matrix_row";
-    else if (field === "role_preference" || field === "event_preference")    type = "multi_select";
-    else if (type === "ignore")                                              type = "string";
-    onChange({ field, type, extra_key: field === "extra_data" ? row.extra_key : "" });
-  }
-
-  function handleTypeChange(type: string) {
-    onChange({ type, ...(type === "ignore" ? { field: "__ignore__" } : {}) });
-  }
-
-  const rowBg = colors
-    ? colors.bg
-    : "var(--color-surface)";
-
-  const borderLeft = colors
-    ? `3px solid ${colors.border}`
-    : "3px solid transparent";
-
-  const opacity = isRemoved ? 0.5 : 1;
+  const isIgnored     = type === "ignore";
+  const needsRowKey   = type === "matrix_row";
+  const needsExtraKey = field === "extra_data";
 
   return (
     <div style={{
       display: "grid", gridTemplateColumns: "1fr 160px 140px 1fr",
       padding: "10px 14px", alignItems: "center", gap: "8px",
-      background: rowBg,
+      background: "var(--color-surface)",
       borderBottom: isLast ? "none" : "1px solid var(--color-border)",
-      borderLeft,
+      borderLeft: "3px solid transparent",
+      opacity: isIgnored ? 0.45 : 1,
     }}>
-      <div style={{ display: "flex", alignItems: "center", gap: "8px", overflow: "hidden" }}>
-        <span style={{
-          fontFamily: "var(--font-mono)", fontSize: "12px",
-          color: "var(--color-text-primary)", opacity,
-          overflow: "hidden", textOverflow: "ellipsis", whiteSpace: "nowrap",
-          textDecoration: isRemoved ? "line-through" : "none",
-        }} title={row.header}>
-          {row.header}
-        </span>
-        {row.state === "new" && (
-          <span style={{ fontFamily: "var(--font-sans)", fontSize: "9px", fontWeight: 700, textTransform: "uppercase", letterSpacing: "0.06em", color: "#16A34A", background: "#DCFCE7", padding: "1px 5px", borderRadius: "3px", flexShrink: 0 }}>
-            New
-          </span>
-        )}
-        {row.state === "removed" && (
-          <span style={{ fontFamily: "var(--font-sans)", fontSize: "9px", fontWeight: 700, textTransform: "uppercase", letterSpacing: "0.06em", color: "#DC2626", background: "#FEE2E2", padding: "1px 5px", borderRadius: "3px", flexShrink: 0 }}>
-            Removed
-          </span>
-        )}
-        {row.state === "changed" && (
-          <span style={{ fontFamily: "var(--font-sans)", fontSize: "9px", fontWeight: 700, textTransform: "uppercase", letterSpacing: "0.06em", color: "#854D0E", background: "#FEF9C3", padding: "1px 5px", borderRadius: "3px", flexShrink: 0 }}>
-            Edited
-          </span>
-        )}
-      </div>
+      {/* Sheet Column — wraps so full text is readable */}
+      <span style={{
+        fontFamily: "var(--font-mono)", fontSize: "12px",
+        color: "var(--color-text-primary)",
+        wordBreak: "break-word",
+        lineHeight: 1.4,
+      }}>
+        {header}
+      </span>
 
-      <select
-        value={row.field}
-        onChange={(e) => handleFieldChange(e.target.value)}
-        disabled={isRemoved}
-        style={{ ...selectStyle, width: "100%", opacity }}
-      >
-        {knownFields.map((f) => (
-          <option key={f} value={f}>{KNOWN_FIELDS_LABELS[f] ?? f}</option>
-        ))}
-      </select>
+      {/* Field */}
+      <span style={{
+        fontFamily: "var(--font-sans)", fontSize: "12px",
+        color: isIgnored ? "var(--color-text-tertiary)" : "var(--color-text-primary)",
+      }}>
+        {isIgnored ? "—" : (KNOWN_FIELDS_LABELS[field] ?? field)}
+      </span>
 
-      <select
-        value={row.type}
-        onChange={(e) => handleTypeChange(e.target.value)}
-        disabled={isRemoved || row.field === "__ignore__"}
-        style={{ ...selectStyle, width: "100%", opacity }}
-      >
-        {validTypes.map((t) => (
-          <option key={t} value={t}>{TYPE_LABELS[t] ?? t}</option>
-        ))}
-      </select>
+      {/* Type */}
+      <span style={{
+        fontFamily: "var(--font-sans)", fontSize: "12px",
+        color: "var(--color-text-secondary)",
+      }}>
+        {TYPE_LABELS[type] ?? type}
+      </span>
 
-      <div style={{ opacity }}>
-        {isRemoved ? (
-          <span style={{ fontFamily: "var(--font-sans)", fontSize: "11px", color: "var(--color-text-tertiary)", fontStyle: "italic" }}>
-            excluded from save
-          </span>
-        ) : needsRowKey ? (
-          <input
-            style={{ ...selectStyle, width: "100%", fontFamily: "var(--font-mono)", fontSize: "11px" }}
-            placeholder="e.g. 8:00 AM - 10:00 AM"
-            value={row.row_key}
-            onChange={(e) => onChange({ row_key: e.target.value })}
-          />
-        ) : needsExtraKey ? (
-          <input
-            style={{ ...selectStyle, width: "100%", fontFamily: "var(--font-mono)", fontSize: "11px" }}
-            placeholder="extra_key name"
-            value={row.extra_key}
-            onChange={(e) => onChange({ extra_key: e.target.value })}
-          />
-        ) : (
-          <span style={{ fontFamily: "var(--font-mono)", fontSize: "11px", color: "var(--color-text-tertiary)" }}>—</span>
-        )}
-      </div>
+      {/* Extra Key / Row Key */}
+      {isIgnored ? (
+        <span style={{ fontFamily: "var(--font-mono)", fontSize: "11px", color: "var(--color-text-tertiary)" }}>—</span>
+      ) : needsRowKey && row_key ? (
+        <span style={{ fontFamily: "var(--font-mono)", fontSize: "11px", color: "var(--color-text-secondary)" }}>{row_key}</span>
+      ) : needsExtraKey && extra_key ? (
+        <span style={{ fontFamily: "var(--font-mono)", fontSize: "11px", color: "var(--color-text-secondary)" }}>{extra_key}</span>
+      ) : (
+        <span style={{ fontFamily: "var(--font-mono)", fontSize: "11px", color: "var(--color-text-tertiary)" }}>—</span>
+      )}
     </div>
   );
 }
 
-// ─── Main Page ────────────────────────────────────────────────────────────────
+// ─── Page ─────────────────────────────────────────────────────────────────────
 
-export default function EditSheetPage() {
+export default function ViewSheetConfigPage() {
   const router = useRouter();
   const params = useParams();
   const tournamentId = Number(params.tournamentId);
   const configId     = Number(params.configId);
 
-  // Config + form fields
-  const [config, setConfig]         = useState<SheetConfig | null>(null);
-  const [label, setLabel]           = useState("");
-  const [sheetType, setSheetType]   = useState("interest");
-  const [selectedTab, setSelectedTab] = useState("");
-  const [availableTabs, setAvailableTabs] = useState<string[]>([]);
-  const [isActive, setIsActive]     = useState(true);
-
-  // Headers + mapping
-  const [mappingRows, setMappingRows]         = useState<RichMappingRow[]>([]);
-  const [knownFields, setKnownFields]         = useState<string[]>([]);
-  const [validTypes, setValidTypes]           = useState<string[]>([]);
-  const [headersLoading, setHeadersLoading]   = useState(false);
-  const [headersError, setHeadersError]       = useState("");
-
-  // Load state
-  const [loading, setLoading]   = useState(true);
+  const [config, setConfig]       = useState<SheetConfig | null>(null);
+  const [loading, setLoading]     = useState(true);
   const [loadError, setLoadError] = useState("");
 
-  // Save / sync
-  const [saveLoading, setSaveLoading]   = useState(false);
-  const [saveError, setSaveError]       = useState("");
-  const [syncLoading, setSyncLoading]   = useState(false);
-  const [saveSuccess, setSaveSuccess]   = useState(false);
-
-  // Delete
   const [showDeleteConfig, setShowDeleteConfig]           = useState(false);
   const [showDeleteMemberships, setShowDeleteMemberships] = useState(false);
   const [deleteLoading, setDeleteLoading]                 = useState(false);
   const [deleteError, setDeleteError]                     = useState("");
-
-  // Import
-  const importInputRef                          = useRef<HTMLInputElement>(null);
-  const [importBanner, setImportBanner]         = useState<{ variant: "success" | "error"; message: string; summary?: ImportSummary } | null>(null);
-  const [showImportSummary, setShowImportSummary] = useState(false);
-
-  // AbortController ref for header fetches
-  const abortRef = useRef<AbortController | null>(null);
-
-  // ── Load config on mount ────────────────────────────────────────────────
 
   useEffect(() => {
     async function load() {
@@ -376,17 +265,6 @@ export default function EditSheetPage() {
       try {
         const cfg = await sheetsApi.getConfig(tournamentId, configId);
         setConfig(cfg);
-        setLabel(cfg.label);
-        setSheetType(cfg.sheet_type);
-        setSelectedTab(cfg.sheet_name);
-        setIsActive(cfg.is_active);
-
-        // Fetch tabs list from Google
-        const validated = await sheetsApi.validate(tournamentId, cfg.sheet_url);
-        setAvailableTabs(validated.sheet_names);
-
-        // Fetch headers for the current tab
-        await fetchHeaders(cfg, cfg.sheet_name);
       } catch {
         setLoadError("Failed to load sheet configuration.");
       } finally {
@@ -394,272 +272,7 @@ export default function EditSheetPage() {
       }
     }
     load();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [tournamentId, configId]);
-
-  // ── Fetch headers (cancellable) ─────────────────────────────────────────
-
-  async function fetchHeaders(cfg: SheetConfig, tabName: string) {
-    // Cancel any in-flight request
-    if (abortRef.current) abortRef.current.abort();
-    const controller = new AbortController();
-    abortRef.current = controller;
-
-    setHeadersLoading(true);
-    setHeadersError("");
-
-    try {
-      const result = await sheetsApi.headers(tournamentId, cfg.sheet_url, tabName);
-
-      // Guard: ignore stale response if a newer request was started
-      if (controller.signal.aborted) return;
-
-      setKnownFields(result.known_fields);
-      setValidTypes(result.valid_types);
-
-      const liveHeaders  = new Set(result.headers);
-      const savedHeaders = new Set(Object.keys(cfg.column_mappings));
-
-      // Build rich rows
-      const rows: RichMappingRow[] = [];
-
-      // Live headers (same or new)
-      for (const header of result.headers) {
-        const saved = cfg.column_mappings[header];
-        if (saved) {
-          // Preexisting header — use saved mapping, state=same initially
-          const row: RichMappingRow = {
-            header,
-            field:     saved.field     ?? "__ignore__",
-            type:      saved.type      ?? "ignore",
-            row_key:   saved.row_key   ?? "",
-            extra_key: saved.extra_key ?? "",
-            state:     "same",
-            original: {
-              header,
-              field:     saved.field     ?? "__ignore__",
-              type:      saved.type      ?? "ignore",
-              row_key:   saved.row_key   ?? "",
-              extra_key: saved.extra_key ?? "",
-            },
-          };
-          rows.push(row);
-        } else {
-          // New header — use auto-detected suggestion
-          const s = result.suggestions[header];
-          const row: RichMappingRow = {
-            header,
-            field:     s?.field     ?? "__ignore__",
-            type:      s?.type      ?? "ignore",
-            row_key:   s?.row_key   ?? "",
-            extra_key: s?.extra_key ?? "",
-            state:     "new",
-            original: {
-              header,
-              field:     s?.field     ?? "__ignore__",
-              type:      s?.type      ?? "ignore",
-              row_key:   s?.row_key   ?? "",
-              extra_key: s?.extra_key ?? "",
-            },
-          };
-          rows.push(row);
-        }
-      }
-
-      // Removed headers — in saved config but not in live sheet
-      for (const header of savedHeaders) {
-        if (!liveHeaders.has(header)) {
-          const saved = cfg.column_mappings[header];
-          rows.push({
-            header,
-            field:     saved.field     ?? "__ignore__",
-            type:      saved.type      ?? "ignore",
-            row_key:   saved.row_key   ?? "",
-            extra_key: saved.extra_key ?? "",
-            state:     "removed",
-            original: {
-              header,
-              field:     saved.field     ?? "__ignore__",
-              type:      saved.type      ?? "ignore",
-              row_key:   saved.row_key   ?? "",
-              extra_key: saved.extra_key ?? "",
-            },
-          });
-        }
-      }
-
-      setMappingRows(rows);
-    } catch (e) {
-      if (controller.signal.aborted) return;
-      setHeadersError("Failed to fetch sheet headers. Check that the sheet is still accessible.");
-    } finally {
-      if (!controller.signal.aborted) setHeadersLoading(false);
-    }
-  }
-
-  // ── Tab change ──────────────────────────────────────────────────────────
-
-  function handleTabChange(tab: string) {
-    setSelectedTab(tab);
-    if (config) fetchHeaders({ ...config, sheet_name: tab }, tab);
-  }
-
-  // ── Update a row + recompute state ──────────────────────────────────────
-
-  function updateRow(idx: number, patch: Partial<MappingRow>) {
-    setMappingRows((prev) =>
-      prev.map((r, i) => {
-        if (i !== idx) return r;
-        const next = { ...r, ...patch };
-        // Recompute state — new stays new, removed stays removed
-        if (r.state === "new" || r.state === "removed") return { ...next, state: r.state };
-        // For preexisting rows, check if anything changed from original
-        const changed =
-          next.field     !== r.original.field     ||
-          next.type      !== r.original.type      ||
-          next.row_key   !== r.original.row_key   ||
-          next.extra_key !== r.original.extra_key;
-        return { ...next, state: changed ? "changed" : "same" };
-      })
-    );
-  }
-
-  // ── Import ──────────────────────────────────────────────────────────────
-
-  function triggerImport(accept: string) {
-    setImportBanner(null);
-    if (importInputRef.current) {
-      importInputRef.current.accept = accept;
-      importInputRef.current.click();
-    }
-  }
-
-  function handleImportFile(e: React.ChangeEvent<HTMLInputElement>) {
-    const file = e.target.files?.[0];
-    if (!file) return;
-    e.target.value = "";
-
-    const isJson = file.name.endsWith(".json") || file.type === "application/json";
-    const isCsv  = file.name.endsWith(".csv")  || file.type === "text/csv";
-
-    const reader = new FileReader();
-    reader.onload = (ev) => {
-      const text = ev.target?.result as string;
-      let parsed: MappingsExport | null = null;
-
-      if (isJson) {
-        parsed = parseMappingsJson(text);
-        if (!parsed) {
-          setImportBanner({ variant: "error", message: "Invalid JSON file — expected { column_mappings: { ... } }" });
-          return;
-        }
-      } else if (isCsv) {
-        parsed = parseMappingsCsv(text);
-        if (!parsed) {
-          setImportBanner({ variant: "error", message: "Invalid CSV file — expected columns: header, field, type, row_key, extra_key" });
-          return;
-        }
-      } else {
-        setImportBanner({ variant: "error", message: "Unsupported file type. Please upload a .json or .csv file." });
-        return;
-      }
-
-      // Only apply to non-removed rows
-      const activeRows: MappingRow[] = mappingRows
-        .filter((r) => r.state !== "removed")
-        .map((r) => ({ header: r.header, field: r.field, type: r.type, row_key: r.row_key, extra_key: r.extra_key }));
-
-      const { updatedRows, summary } = applyImport(activeRows, parsed);
-
-      // Merge back into rich rows, preserving state logic
-      setMappingRows((prev) =>
-        prev.map((r) => {
-          if (r.state === "removed") return r;
-          const updated = updatedRows.find((u) => u.header === r.header);
-          if (!updated) return r;
-          const changed =
-            updated.field     !== r.original.field     ||
-            updated.type      !== r.original.type      ||
-            updated.row_key   !== r.original.row_key   ||
-            updated.extra_key !== r.original.extra_key;
-          return {
-            ...r,
-            ...updated,
-            state: r.state === "new" ? "new" : changed ? "changed" : "same",
-          };
-        })
-      );
-
-      if (parsed.label && !label) setLabel(parsed.label);
-      if (parsed.sheet_type) setSheetType(parsed.sheet_type);
-
-      const { updated: updatedList, unchanged, notInSheet, notInFile } = summary;
-      const shortMsg = `${updatedList.length} updated, ${unchanged} unchanged, ${notInSheet.length} ignored, ${notInFile.length} untouched`;
-      setImportBanner({ variant: "success", message: `Import successful: ${shortMsg}`, summary });
-    };
-    reader.readAsText(file);
-  }
-
-  // ── Build column mappings for save (exclude removed rows) ───────────────
-
-  const buildColumnMappings = useCallback((): Record<string, ColumnMapping> => {
-    const result: Record<string, ColumnMapping> = {};
-    for (const row of mappingRows) {
-      if (row.state === "removed") continue;
-      const mapping: ColumnMapping = { field: row.field, type: row.type as ColumnMapping["type"] };
-      if (row.type === "matrix_row" && row.row_key) mapping.row_key = row.row_key;
-      if (row.field === "extra_data" && row.extra_key) mapping.extra_key = row.extra_key;
-      result[row.header] = mapping;
-    }
-    return result;
-  }, [mappingRows]);
-
-  // ── Save ────────────────────────────────────────────────────────────────
-
-  async function handleSave() {
-    setSaveLoading(true);
-    setSaveError("");
-    setSaveSuccess(false);
-    try {
-      await sheetsApi.updateConfig(tournamentId, configId, {
-        label,
-        sheet_type: sheetType as SheetConfig["sheet_type"],
-        sheet_name: selectedTab,
-        column_mappings: buildColumnMappings(),
-        is_active: isActive,
-      });
-      setSaveSuccess(true);
-    } catch {
-      setSaveError("Failed to save changes.");
-    } finally {
-      setSaveLoading(false);
-    }
-  }
-
-  // ── Save & Sync ─────────────────────────────────────────────────────────
-
-  async function handleSaveAndSync() {
-    setSyncLoading(true);
-    setSaveError("");
-    setSaveSuccess(false);
-    try {
-      await sheetsApi.updateConfig(tournamentId, configId, {
-        label,
-        sheet_type: sheetType as SheetConfig["sheet_type"],
-        sheet_name: selectedTab,
-        column_mappings: buildColumnMappings(),
-        is_active: isActive,
-      });
-      await sheetsApi.sync(tournamentId, configId);
-      setSaveSuccess(true);
-    } catch {
-      setSaveError("Failed to save or sync.");
-    } finally {
-      setSyncLoading(false);
-    }
-  }
-
-  // ── Delete config only ──────────────────────────────────────────────────
 
   async function handleDeleteConfig() {
     setDeleteLoading(true);
@@ -673,16 +286,11 @@ export default function EditSheetPage() {
     }
   }
 
-  // ── Delete config + memberships ─────────────────────────────────────────
-
   async function handleDeleteMemberships() {
     if (!config) return;
     setDeleteLoading(true);
     setDeleteError("");
     try {
-      // Get email list from existing memberships in this tournament
-      // TEMP: does not cross-reference live sheet — deletes all memberships in the tournament.
-      // TODO: replace with sheetsApi.getRows once the backend endpoint exists.
       let emails: string[];
       try {
         emails = await sheetsApi.getEmailsForNuclearDelete(tournamentId);
@@ -692,10 +300,7 @@ export default function EditSheetPage() {
         setShowDeleteMemberships(false);
         return;
       }
-
-      // Delete memberships by email
       await membershipsApi.deleteMembershipsByEmails(tournamentId, emails);
-      // Delete the config
       await sheetsApi.deleteConfig(tournamentId, configId);
       router.push(`/dashboard/${tournamentId}/sheets`);
     } catch {
@@ -704,13 +309,6 @@ export default function EditSheetPage() {
       setShowDeleteMemberships(false);
     }
   }
-
-  // ── Summary counts for header banner ───────────────────────────────────
-
-  const sameCount    = mappingRows.filter((r) => r.state === "same").length;
-  const newCount     = mappingRows.filter((r) => r.state === "new").length;
-  const removedCount = mappingRows.filter((r) => r.state === "removed").length;
-  const changedCount = mappingRows.filter((r) => r.state === "changed").length;
 
   // ── Render ────────────────────────────────────────────────────────────────
 
@@ -723,142 +321,98 @@ export default function EditSheetPage() {
     );
   }
 
-  if (loadError) {
+  if (loadError || !config) {
     return (
       <div style={{ width: "100%" }}>
         <BackLink onClick={() => router.push(`/dashboard/${tournamentId}/sheets`)} />
-        <p style={{ fontFamily: "var(--font-sans)", fontSize: "13px", color: "var(--color-danger)" }}>{loadError}</p>
+        <p style={{ fontFamily: "var(--font-sans)", fontSize: "13px", color: "var(--color-danger)" }}>
+          {loadError || "Config not found."}
+        </p>
       </div>
     );
   }
 
+  const mappingEntries = Object.entries(config.column_mappings);
+
   return (
     <div style={{ width: "100%" }}>
-      <BackLink onClick={() => router.push(`/dashboard/${tournamentId}/sheets`)} />
 
-      <h1 style={{ fontSize: "28px", lineHeight: 1.2, marginBottom: "4px" }}>Edit Sheet Config</h1>
-      <p style={{ fontFamily: "var(--font-sans)", fontSize: "13px", color: "var(--color-text-secondary)", marginBottom: "28px" }}>
-        {config?.sheet_url && (
-          <span style={{ fontFamily: "var(--font-mono)", fontSize: "11px", color: "var(--color-text-tertiary)" }}>
-            {config.sheet_url}
-          </span>
-        )}
+      {/* ── Header ── */}
+      <div style={{ display: "flex", alignItems: "flex-start", justifyContent: "space-between" }}>
+        <BackLink onClick={() => router.push(`/dashboard/${tournamentId}/sheets`)} />
+        <Button
+          variant="secondary"
+          size="sm"
+          onClick={() => router.push(`/dashboard/${tournamentId}/sheets/${configId}/edit`)}
+        >
+          <IconEdit size={13} />
+          Edit
+        </Button>
+      </div>
+
+      <h1 style={{ fontSize: "28px", lineHeight: 1.2, marginBottom: "4px" }}>{config.label}</h1>
+      <p style={{ fontFamily: "var(--font-mono)", fontSize: "11px", color: "var(--color-text-tertiary)", marginBottom: "28px" }}>
+        {config.sheet_url}
       </p>
 
       <div style={{ display: "flex", flexDirection: "column", gap: "24px" }}>
 
-        {/* ── Config fields ── */}
-        <div style={{ display: "grid", gridTemplateColumns: "1fr 1fr", gap: "16px" }}>
-          <div>
-            <FieldLabel htmlFor="label">Label</FieldLabel>
-            <input
-              id="label"
-              value={label}
-              onChange={(e) => setLabel(e.target.value)}
-              style={{ ...selectStyle, width: "100%", height: "44px", padding: "0 14px" }}
-            />
-          </div>
-          <div>
-            <FieldLabel>Sheet Type</FieldLabel>
-            <select value={sheetType} onChange={(e) => setSheetType(e.target.value)} style={{ ...selectStyle, width: "100%", height: "44px" }}>
-              {SHEET_TYPES.map(({ value, label: l }) => (
-                <option key={value} value={value}>{l}</option>
-              ))}
-            </select>
-          </div>
-        </div>
-
-        <div style={{ display: "grid", gridTemplateColumns: "1fr auto", gap: "16px", alignItems: "end" }}>
-          <div>
-            <FieldLabel>Sheet Tab</FieldLabel>
-            <select value={selectedTab} onChange={(e) => handleTabChange(e.target.value)} style={{ ...selectStyle, width: "100%", height: "44px" }}>
-              {availableTabs.map((tab) => (
-                <option key={tab} value={tab}>{tab}</option>
-              ))}
-            </select>
-          </div>
-          <label style={{ display: "flex", alignItems: "center", gap: "8px", height: "44px", cursor: "pointer", fontFamily: "var(--font-sans)", fontSize: "13px", color: "var(--color-text-primary)", flexShrink: 0 }}>
-            <input
-              type="checkbox"
-              checked={isActive}
-              onChange={(e) => setIsActive(e.target.checked)}
-              style={{ accentColor: "var(--color-accent)", width: "14px", height: "14px" }}
-            />
-            Active
-          </label>
+        {/* ── Metadata — single row, 4 equal sections ── */}
+        <div style={{
+          display: "grid",
+          gridTemplateColumns: "1fr 1fr 1fr 1fr",
+          border: "1px solid var(--color-border)",
+          borderRadius: "var(--radius-md)",
+          overflow: "hidden",
+        }}>
+          {[
+            { label: "Type",       value: SHEET_TYPE_LABELS[config.sheet_type] ?? config.sheet_type },
+            { label: "Sheet Tab",  value: config.sheet_name },
+            { label: "Status",     value: config.is_active ? "Active" : "Inactive" },
+            { label: "Last Synced", value: config.last_synced_at ? fmtDateTime(config.last_synced_at) : "Never" },
+          ].map(({ label, value }, i, arr) => (
+            <div
+              key={label}
+              style={{
+                padding: "14px 18px",
+                borderRight: i < arr.length - 1 ? "1px solid var(--color-border)" : "none",
+                background: "var(--color-surface)",
+              }}
+            >
+              <p style={{
+                fontFamily: "var(--font-sans)", fontSize: "10px", fontWeight: 600,
+                textTransform: "uppercase", letterSpacing: "0.07em",
+                color: "var(--color-text-tertiary)", marginBottom: "4px",
+              }}>
+                {label}
+              </p>
+              <p style={{
+                fontFamily: "var(--font-mono)", fontSize: "12px",
+                color: "var(--color-text-primary)",
+                whiteSpace: "nowrap", overflow: "hidden", textOverflow: "ellipsis",
+              }}>
+                {value}
+              </p>
+            </div>
+          ))}
         </div>
 
         {/* ── Mapping table ── */}
         <div>
-          {/* Toolbar */}
-          <div style={{ display: "flex", alignItems: "center", justifyContent: "space-between", gap: "12px", marginBottom: "10px" }}>
-            {/* Summary counts */}
-            {!headersLoading && mappingRows.length > 0 && (
-              <div style={{ display: "flex", gap: "12px", flexWrap: "wrap" }}>
-                {[
-                  { label: `${sameCount} unchanged`, color: "var(--color-text-tertiary)" },
-                  ...(changedCount > 0 ? [{ label: `${changedCount} edited`, color: "#854D0E" }] : []),
-                  ...(newCount > 0     ? [{ label: `${newCount} new`,        color: "#16A34A" }] : []),
-                  ...(removedCount > 0 ? [{ label: `${removedCount} removed`, color: "#DC2626" }] : []),
-                ].map(({ label: l, color }) => (
-                  <span key={l} style={{ fontFamily: "var(--font-sans)", fontSize: "12px", color }}>{l}</span>
-                ))}
-              </div>
-            )}
-            {headersLoading && (
-              <span style={{ fontFamily: "var(--font-sans)", fontSize: "12px", color: "var(--color-text-tertiary)" }}>
-                Fetching headers…
-              </span>
-            )}
-
-            {/* Import split button */}
-            <div style={{ flexShrink: 0 }}>
-              <SplitButton
-                label="Import JSON"
-                onClick={() => triggerImport(".json,application/json")}
-                variant="secondary"
-                size="sm"
-                options={[
-                  { label: "Import JSON", action: () => triggerImport(".json,application/json") },
-                  { label: "Import CSV",  action: () => triggerImport(".csv,text/csv") },
-                ]}
-              />
-              <input
-                ref={importInputRef}
-                type="file"
-                accept=".json,.csv,application/json,text/csv"
-                style={{ display: "none" }}
-                onChange={handleImportFile}
-              />
-            </div>
+          <div style={{ marginBottom: "10px" }}>
+            <span style={{ fontFamily: "var(--font-sans)", fontSize: "12px", color: "var(--color-text-tertiary)" }}>
+              {mappingEntries.filter(([, m]) => m.type !== "ignore").length} mapped,{" "}
+              {mappingEntries.filter(([, m]) => m.type === "ignore").length} ignored
+            </span>
           </div>
 
-          {/* Import banner */}
-          {importBanner && (
-            <div style={{ marginBottom: "10px" }}>
-              <Banner
-                variant={importBanner.variant}
-                message={importBanner.message}
-                onDismiss={() => setImportBanner(null)}
-                action={importBanner.summary ? (
-                  <Button variant="ghost" size="sm" onClick={() => setShowImportSummary(true)}>
-                    Show summary
-                  </Button>
-                ) : undefined}
-              />
-            </div>
-          )}
-
-          {headersError && (
-            <p style={{ fontFamily: "var(--font-sans)", fontSize: "12px", color: "var(--color-danger)", marginBottom: "10px" }}>
-              {headersError}
+          {mappingEntries.length === 0 ? (
+            <p style={{ fontFamily: "var(--font-sans)", fontSize: "13px", color: "var(--color-text-tertiary)" }}>
+              No columns mapped yet.
             </p>
-          )}
-
-          {/* Table */}
-          {!headersLoading && mappingRows.length > 0 && (
+          ) : (
             <div style={{ border: "1px solid var(--color-border)", borderRadius: "var(--radius-md)", overflow: "hidden" }}>
-              {/* Header row */}
+              {/* Column headers */}
               <div style={{
                 display: "grid", gridTemplateColumns: "1fr 160px 140px 1fr",
                 padding: "8px 14px",
@@ -866,45 +420,29 @@ export default function EditSheetPage() {
                 borderBottom: "1px solid var(--color-border)",
               }}>
                 {["Sheet Column", "Field", "Type", "Extra Key / Row Key"].map((h) => (
-                  <span key={h} style={{ fontFamily: "var(--font-sans)", fontSize: "10px", fontWeight: 600, textTransform: "uppercase", letterSpacing: "0.07em", color: "var(--color-text-tertiary)" }}>
+                  <span key={h} style={{
+                    fontFamily: "var(--font-sans)", fontSize: "10px", fontWeight: 600,
+                    textTransform: "uppercase", letterSpacing: "0.07em",
+                    color: "var(--color-text-tertiary)",
+                  }}>
                     {h}
                   </span>
                 ))}
               </div>
 
-              {mappingRows.map((row, idx) => (
-                <MappingTableRow
-                  key={row.header}
-                  row={row}
-                  knownFields={knownFields}
-                  validTypes={validTypes}
-                  onChange={(patch) => updateRow(idx, patch)}
-                  isLast={idx === mappingRows.length - 1}
+              {mappingEntries.map(([header, mapping], idx) => (
+                <ReadOnlyMappingRow
+                  key={header}
+                  header={header}
+                  field={mapping.field}
+                  type={mapping.type}
+                  row_key={mapping.row_key}
+                  extra_key={mapping.extra_key}
+                  isLast={idx === mappingEntries.length - 1}
                 />
               ))}
             </div>
           )}
-        </div>
-
-        {/* ── Save actions ── */}
-        <div style={{ display: "flex", flexDirection: "column", gap: "10px" }}>
-          {saveError && (
-            <p style={{ fontFamily: "var(--font-sans)", fontSize: "12px", color: "var(--color-danger)" }}>{saveError}</p>
-          )}
-          {saveSuccess && (
-            <Banner variant="success" message="Changes saved successfully." onDismiss={() => setSaveSuccess(false)} />
-          )}
-          <div style={{ display: "flex", gap: "10px" }}>
-            <Button variant="secondary" size="md" onClick={() => router.push(`/dashboard/${tournamentId}/sheets`)}>
-              Cancel
-            </Button>
-            <Button variant="secondary" size="md" loading={saveLoading} onClick={handleSave}>
-              Save
-            </Button>
-            <Button variant="primary" size="md" loading={syncLoading} onClick={handleSaveAndSync}>
-              Save &amp; Sync
-            </Button>
-          </div>
         </div>
 
         {/* ── Danger zone ── */}
@@ -938,32 +476,27 @@ export default function EditSheetPage() {
             </Button>
           </div>
           <p style={{ fontFamily: "var(--font-sans)", fontSize: "12px", color: "var(--color-text-tertiary)" }}>
-            "Delete config" removes this configuration only — volunteer data is unaffected.
-            "Delete config + memberships" also removes all memberships in this tournament whose email appears in the current sheet.
+            &quot;Delete config&quot; removes this configuration only — volunteer data is unaffected.
+            &quot;Delete config + memberships&quot; also removes all memberships in this tournament whose email appears in the current sheet.
           </p>
         </div>
+
       </div>
 
       {/* ── Modals ── */}
-      {showDeleteConfig && config && (
+      {showDeleteConfig && (
         <DeleteConfigModal
           label={config.label}
           onConfirm={handleDeleteConfig}
           onCancel={() => setShowDeleteConfig(false)}
         />
       )}
-      {showDeleteMemberships && config && (
+      {showDeleteMemberships && (
         <DeleteMembershipsModal
           label={config.label}
           onConfirm={handleDeleteMemberships}
           onCancel={() => setShowDeleteMemberships(false)}
           loading={deleteLoading}
-        />
-      )}
-      {showImportSummary && importBanner?.summary && (
-        <ImportSummaryModal
-          summary={importBanner.summary}
-          onClose={() => setShowImportSummary(false)}
         />
       )}
     </div>
